Fixed bug with incorrect negatives
When creating amount owed to someone, DB should store as negative not
positive number. List view should show amounts < 0.
When creating money owed to you, DB should store as positive not
negative number. List view should show amounts > 0

This corresponds to the way it is displayed on the main screen